fix win compiler warnings
git-svn-id: https://simulationcraft.googlecode.com/svn/trunk@1044 3b4652a1-8050-0410-ac47-3d40261b0f8b

+++ sc_sim.cpp
@@ -203,7 +203,7 @@ bool sim_t::init()
 {
   total_seconds = 0;
 
-  if( seed == 0 ) seed = time( NULL );
+  if( seed == 0 ) seed = (int32_t) time( NULL );
   rand_t::init( seed );
 
   if( ! patch_str.empty() )
@@ -528,7 +528,7 @@ void sim_t::execute()
   merge();
   analyze();
 
-  elapsed_cpu_seconds = time(0) - start_time;
+  elapsed_cpu_seconds = (double) ( time(0) - start_time );
 }
 
 // sim_t::find_player =======================================================
